Fix: rename 'Maddening Touch.tsx' to 'MaddeningTouch.tsx'
All other talent module files under the shadow priest directory use
PascalCase filenames without spaces (e.g. SurgeOfInsanity.tsx,
ShatteredPsyche.tsx). The file 'Maddening Touch.tsx' was the sole
exception, which is inconsistent and can cause issues with tools that
do not handle spaces in module paths well.

Rename the file to MaddeningTouch.tsx and update the corresponding
import in CombatLogParser.tsx.
