+++ src/analysis/retail/priest/shadow/CombatLogParser.tsx
@@ -48,7 +48,7 @@ import PsychicLink from './modules/talents/PsychicLink';
 import VoidTouched from './modules/talents/VoidTouched';
 import MindsEye from './modules/talents/MindsEye';
 import DistortedReality from './modules/talents/DistortedReality';
-import MaddeningTouch from './modules/talents/Maddening Touch';
+import MaddeningTouch from './modules/talents/MaddeningTouch';
 import ShatteredPsyche from './modules/talents/ShatteredPsyche';
 import Mastermind from './modules/talents/Mastermind';
 import DarkAscension from './modules/talents/DarkAscension';
